+++ dss-apps/dss-dataasset-management/data-assets-server/src/main/java/com/webank/wedatasphere/dss/data/governance/service/AssetService.java
@@ -17,7 +17,7 @@ import java.util.Set;
 
 
 public interface AssetService {
-    public Map<String,Integer> getHiveSummary() throws DataGovernanceException;
+    public Map<String,Long> getHiveSummary() throws DataGovernanceException;
 
     public List<HiveTblSimpleInfo> searchHiveTable(String classification, String query,String termName,
                                                    int limit, int offset) throws DataGovernanceException;
